// components/CConditions/interface.ts
export interface CRelationProps {
  /**
   * @zh 尺寸大小
   * @en size
   * @defaultValue 'default'
   */
  size?: 'small' | 'default';
  /**
   * @zh 是否是展示态（默认为否，可编辑）
   * @en Display mode
   * @defaultValue false
   */
  readonly?: boolean;
  /**
   * @zh 是否禁止修改
   * @en Disable editing
   * @defaultValue false
   */
  disabled?: boolean;
  /**
   * @zh 值
   * @en Value
   */
  value?: string | number;
  /**
   * @zh 值改变时的回调
   * @en Callback when value changes
   */
  onChange?: (value: string | number) => void;
  /**
   * @zh 逻辑关系的选项
   * @en Logic relation options
   * @defaultValue ['and', 'or']
   */
  options?: SelectProps['options'];
  /**
   * @zh 自定义选择器，`options` 和 `children` 二选一即可，但 `options` 优先级更高
   * @en Custom selector, `options` and `children` can be selected, but `options` has a higher priority
   */
  children?: ReactNode;
  style?: CSSProperties;
  className?: string;
}